requirements.txt
fastapi>=0.108.0
uvicorn>=0.25.0
ultralytics>=8.0.216
torch>=2.1.2
torchvision>=0.16.2
opencv-python>=4.8.1.78
numpy>=1.24.3
Pillow>=10.1.0
pydantic>=2.5.3
python-multipart>=0.0.6